feat: Add jito fee service and simulation endpoint
This commit introduces the `jito_fee_service` and adds a new
`simulateBundle` endpoint to the `bundleService`.

Key changes include:
- Adding `jito_fee_service` to the Dockerfile and build process.
- Implementing the `simulateBundle` RPC method in the `bundleService`.
- Adding a new `TransactionSimulationDetails` component to display
  simulation results in the UI.
- Updating the `TransactionItem` to utilize the new simulation details.
- Adding a new test file `simulateBundle.test.ts` to verify the
  functionality of the `simulateBundle` endpoint.
- Removing unused `atl_redis_conn` function and related configuration.
- Updating `jupiter_trader_data` to use `f64` for route plan percent.
- Updating `reqwest_client` to use `anyhow::anyhow` for error handling.
- Adjusting `txId` to `id` in `TransactionSimulationDetails` for
  consistency.

+++ jupiter_trader_data/src/models/jupiter_models.rs
@@ -306,7 +306,7 @@ pub struct PriorityLevelWithMaxLamports {
 pub struct RoutePlan {
     #[serde(rename = "swapInfo")]
     pub swap_info: SwapInfo,
-    pub percent: u8,
+    pub percent: f64,
     pub bps: Option<u64>,
     #[serde(rename = "usdValue")]
     pub usd_value: Option<String>,
@@ -316,7 +316,7 @@ pub struct RoutePlan {
 pub struct RoutePlanUltra {
     #[serde(rename = "swapInfo")]
     pub swap_info: SwapInfo,
-    pub percent: u8,
+    pub percent: f64,
     pub bps: u64,
 }
